TangencyViaPoint - automatic point-on-object+bugfixes
Point-on-object constraints are added automatically in UI handler.
Point-on-curve are not autoadded when the point is an endpoint of the
curve (otherwise, it still generates a redundant constraint if the point
was constraint to one of the curves due to autoconstraints)
The point-on-object constraints are kept as separate UI constraints for
being able to manually resolve arising conflicts, and for compatibility
with potential future changes to automatic additions of point-on-object
(old files will still load regardless).