Base: UnsafePyHandle implementation using py::object composition
================================================================

TODO: This commit compiles, but crashes with example:
https://forum.freecadweb.org/viewtopic.php?f=20&t=40525&start=190#p379164

so a fix is needed.

TODO: Move implementation to cpp file

This is a rewrite of DeepSOIC's UnsafePyHandle using composition instead of public inheritance.

What exactly and why?

In the original branch UnsafePyHandle derived from Py::Object. This allowed implicitly assigning almost anything to anything.

The first design decision is to use composition instead of deriving from Py::Object (with private inheritance it would have worked, but the compilation error messages were not meaningful, as for the compiler the problem was the inheritance being private, not the actual underlying issue).

The second is to make pyobject* and py::object constructors explicit to avoid using them inadvertently.

The third is providing for "implicit" upcast between UnsafePyHandle<Derived> and UnsafePyHandle<Base> with compilation time type checking. These functions are not optimized and a copy is currently done. I have thought of using moving syntax, but being pointers involved it may not be worth it. I have thought of separating copy constructor from the upcast and implement upcast using reinterpret_ptr. Let's see.

The fourth is requiring explicit downcast between UnsafePyHandle<Derived> and UnsafePyHandle<Base> with compilation time type checking.

The fifth is not to allow copy assignment from Py::Object (only construction is possible) to enforce type checking. A new method has been created "makeNone()" if an assignment of Py::None is necessary, that takes care to making the soviet pointer cppptr to nullptr at the same time.

The sixth is to expose a minimum Py::Object interface not to change too much code in this iteration. I have to review that all is needed in a near future.

The seventh is that I have maintained (or at least I have tried to) the original way of setting the pointer, but with the method now called set() instead of setObject(), but I have split checks from exception throwing. I do not think it is acceptable that a user of a class can make the program throw just by checking a type (another thing is for assigning what he should not assign).

The eight is that when using composition, the protected functions of py::object are not available, this means that set with owned=true is emulated by decreasing the reference counter after assignment via the operator=() of py::object.

Why not using private inheritance to avoid the incrementing the reference only to decrement it afterwards in eight?

Because when using private inheritance the compiler errors of the code that uses UnsafePyHandle tend not to show the real problem, as the compiler things is everything about Py::Object data members not being accessible due to the private inheritance. This hides the real problem and makes bug fixing very cumbersome. It is indeed an option (the first I tried in a previous iteration and is viable other than by the compiler errors).