[CONFLICT] Sketcher: SketchSolver interface
================================

This is a base class for solver interfaces (Sketch.h for GCS).

It defines the minimum interface for sketcher->solver interface interaction.

This build removes:
-> Advanced solver dialog
-> Ability to retrieve non-constraint geometry (missing dofs)

The reason is that there is a strong coupling between the specific solver GCS and these functionality above.

A better interface needs to be rewritten.